+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 ------------
 
+##### 7.1.0
+Added `scrollToAlignment` property to `Collection`, `Grid`, `FlexTable`, and `VirtualScroll` to offer finer-grained control of how scrolled-to cells are aligned.
+Default behavior ("_auto_") remains unchanged- the least amount of scrolling will occur to ensure that the specified cell is visible.
+
 ##### 7.0.5
 Fixed edge-case bug where `InfiniteLoader` did not respect `minBatchSize` setting when a user was scrolling up.